requirements.txt
appier==1.30.0
beautifulsoup4==4.11.1
bs4==0.0.1
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
flake8==6.0.0
google-api==0.1.12
google-api-core==2.10.2
google-api-python-client==2.65.0
google-auth==2.14.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.7.1
googleapis-common-protos==1.56.4
httplib2==0.21.0
idna==3.4
isort==5.10.1
lxml==4.9.2
mccabe==0.7.0
oauthlib==3.2.2
oauth2client==3.0.0
pathmatch==0.2.2
protobuf==4.21.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pyflakes==3.0.1
pyparsing==3.0.9
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
soupsieve==2.3.2.post1
typing==3.7.4.3
uritemplate==4.1.1
urllib3==1.26.12